+++ packages/cli/src/handlers/generate.manifest.ts
@@ -13,18 +13,23 @@ export type GenerateManifestArgs = {
 export default (argv: GenerateManifestArgs) => {
   const { configsDir, depth, outDir, outFile } = argv;
 
-  console.log(
-    `generating manifest --configsDir=${configsDir} --outDir=${outDir} --outFile=${outFile}`,
-  );
+  try {
+    console.log(
+      `generating manifest --configsDir=${configsDir} --outDir=${outDir} --outFile=${outFile}`,
+    );
 
-  if (!fs.existsSync(configsDir)) {
-    console.log(`configsDir not found '${configsDir}'. skipping manifest generation.`);
-  } else {
-    const config = generate(configsDir, depth);
+    if (!fs.existsSync(configsDir)) {
+      console.log(`configsDir not found '${configsDir}'. skipping manifest generation.`);
+    } else {
+      const config = generate(configsDir, depth);
 
-    if (!fs.existsSync(outDir)) {
-      fs.mkdirSync(outDir);
+      if (!fs.existsSync(outDir)) {
+        fs.mkdirSync(outDir);
+      }
+      fs.writeFileSync(path.join(outDir, outFile), JSON.stringify(config));
     }
-    fs.writeFileSync(path.join(outDir, outFile), JSON.stringify(config));
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  } catch (err: any) {
+    console.error('Error generating manifest', err.message);
   }
 };