+++ packages/react-host/CHANGELOG.md
@@ -3,6 +3,64 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.3](https://github.com/navaris/appshell/compare/@appshell/react-host@0.1.0-alpha.2...@appshell/react-host@0.1.0-alpha.3) (2023-06-28)
+
+
+### Features
+
+* enable props to be passed to federated component rendered by react host ([e3a7273](https://github.com/navaris/appshell/commit/e3a72738455e179c9cc0b3f1ae29cf61b7c6d765))
+
+
+
+
+
+# [0.1.0-alpha.2](https://github.com/navaris/appshell/compare/@appshell/react-host@0.1.0-alpha.1...@appshell/react-host@0.1.0-alpha.2) (2023-06-26)
+
+
+### Bug Fixes
+
+* ignore dist from dev server ([66eaa4f](https://github.com/navaris/appshell/commit/66eaa4f0f7909de289aa25bd8962d0d018be8c89))
+
+
+
+
+
+# [0.1.0-alpha.1](https://github.com/navaris/appshell/compare/@appshell/react-host@0.1.0-alpha.0...@appshell/react-host@0.1.0-alpha.1) (2023-06-26)
+
+
+### Bug Fixes
+
+* ignore manifest changes that force refresh ([ccb2155](https://github.com/navaris/appshell/commit/ccb2155a26bd3f6861f43805bb674ae4906375c1))
+
+
+
+
+
+# [0.1.0-alpha.0](https://github.com/navaris/appshell/compare/@appshell/react-host@0.0.2-alpha.0...@appshell/react-host@0.1.0-alpha.0) (2023-06-22)
+
+
+### Features
+
+* support manifest merging and config watch ([5431d10](https://github.com/navaris/appshell/commit/5431d100ec7f5106cab66d3009ce5f836e452715))
+
+
+
+
+
+## [0.0.2-alpha.0](https://github.com/navaris/appshell/compare/@appshell/react-host@0.0.1...@appshell/react-host@0.0.2-alpha.0) (2023-06-15)
+
+
+### Bug Fixes
+
+* enable history api fallback, include serve.json config, update publicPath value ([96af4fc](https://github.com/navaris/appshell/commit/96af4fc8cd14006aa477964daa4e37d29803191c))
+* enable relative appshell_configs dir, containerize appshell host example ([79926f5](https://github.com/navaris/appshell/commit/79926f5e2b1ae29c4d7349c0babcb93d3b147485))
+* hot module replacement support, example updates ([a67d230](https://github.com/navaris/appshell/commit/a67d2303b40f5911373225cc0b4ccf9b67c33d11))
+* make runtime env generation prefix configurable ([28441a8](https://github.com/navaris/appshell/commit/28441a854cd03dfd4b89fb932efba571af221cd2))
+
+
+
+
+
 ## 0.0.1 (2023-06-15)
 
 **Note:** Version bump only for package @appshell/react-host